refactor: join data module
