+++ app/build.gradle.kts
@@ -73,8 +73,6 @@ android {
 dependencies {
     implementation(project(":core:model"))
     implementation(project(":core:data"))
-    implementation(project(":core:database"))
-    implementation(project(":core:datastore"))
     implementation(project(":core:domain"))
     implementation(project(":core:resource"))
     implementation(project(":core:ui"))
