+++ core/data/build.gradle.kts
@@ -3,6 +3,7 @@ plugins {
     id(libs.plugins.hilt.get().pluginId)
     id(libs.plugins.kotlin.android.get().pluginId)
     id(libs.plugins.kotlin.kapt.get().pluginId)
+    id(libs.plugins.ksp.get().pluginId)
 }
 
 android {
@@ -10,6 +11,11 @@ android {
     namespace = "kosenda.makecolor.core.data"
     defaultConfig {
         minSdk = 24
+        javaCompileOptions {
+            annotationProcessorOptions {
+                arguments += mapOf("room.schemaLocation" to "$projectDir/schemas")
+            }
+        }
     }
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_11
@@ -19,11 +25,13 @@ android {
 
 dependencies {
     implementation(project(":core:model"))
-    implementation(project(":core:database"))
-    implementation(project(":core:datastore"))
     implementation(project(":core:resource"))
+    implementation(libs.androidx.dataStore.preferences)
     implementation(libs.hilt.android)
+    implementation(libs.kotlinx.serialization.json)
     implementation(libs.room.runtime)
+    implementation(libs.room.ktx)
     implementation(libs.timber)
     kapt(libs.hilt.compiler)
+    ksp(libs.room.compiler)
 }
