+++ core/data/src/main/java/kosenda/makecolor/core/data/database/ColorDatabase.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor.core.database
+package kosenda.makecolor.core.data.database
 
 import android.content.Context
 import androidx.room.Database
