+++ core/data/src/main/java/kosenda/makecolor/core/data/di/ColorDaoModule.kt
@@ -1,11 +1,11 @@
-package kosenda.makecolor.core.database.di
+package kosenda.makecolor.core.data.di
 
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.components.SingletonComponent
-import kosenda.makecolor.core.database.ColorDao
-import kosenda.makecolor.core.database.ColorDatabase
+import kosenda.makecolor.core.data.database.ColorDao
+import kosenda.makecolor.core.data.database.ColorDatabase
 import javax.inject.Singleton
 
 @Module
