+++ core/data/src/main/java/kosenda/makecolor/core/data/di/ColorDatabaseModule.kt
@@ -1,4 +1,4 @@
-package kosenda.makecolor.core.database.di
+package kosenda.makecolor.core.data.di
 
 import android.content.Context
 import dagger.Module
@@ -6,7 +6,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import kosenda.makecolor.core.database.ColorDatabase
+import kosenda.makecolor.core.data.database.ColorDatabase
 import javax.inject.Singleton
 
 @Module
