--- core/database/build.gradle.kts
@@ -1,34 +1,0 @@
-plugins {
-    id(libs.plugins.android.library.get().pluginId)
-    id(libs.plugins.hilt.get().pluginId)
-    id(libs.plugins.kotlin.android.get().pluginId)
-    id(libs.plugins.kotlin.kapt.get().pluginId)
-    id(libs.plugins.ksp.get().pluginId)
-}
-
-android {
-    compileSdk = 33
-    namespace = "kosenda.makecolor.core.database"
-    defaultConfig {
-        minSdk = 24
-        javaCompileOptions {
-            annotationProcessorOptions {
-                arguments += mapOf("room.schemaLocation" to "$projectDir/schemas")
-            }
-        }
-    }
-    compileOptions {
-        sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
-    }
-}
-
-dependencies {
-    implementation(project(":core:model"))
-    implementation(libs.hilt.android)
-    implementation(libs.kotlinx.serialization.json)
-    implementation(libs.room.runtime)
-    implementation(libs.room.ktx)
-    kapt(libs.hilt.compiler)
-    ksp(libs.room.compiler)
-}
