+++ core/ui/build.gradle.kts
@@ -24,7 +24,6 @@ android {
 
 dependencies {
     implementation(project(":core:data"))
-    implementation(project(":core:database"))
     implementation(project(":core:model"))
     implementation(project(":core:resource"))
     implementation(project(":core:util"))
